MINOR: Update nagios runit to verify configuration on start
We want to stick as much as possible to what is done in equivalent
systemd unit file. That script update-nagios in Kompot, when a new
configuration fails, hints the user about the error found in
journal/syslog logs. On systemd installation this works because
the unit verifies nagios configuration with an ExecStartPre and
the output gets eaten by journal/syslog. Let's add that same
verification in nagios runit run script, making sure its output
gets logged to syslog.